pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "kinetic-core"
version = "1.0.0"
description = "Autonomous Reliability Engineer for Critical Power Systems"
requires-python = ">=3.11"
dependencies = [
    # Azure AI / OpenAI
    "openai>=1.30.0",
    "azure-ai-textanalytics>=5.3.0",
    # Azure Storage & Cosmos
    "azure-cosmos>=4.7.0",
    "azure-storage-blob>=12.20.0",
    # Azure Search
    "azure-search-documents>=11.6.0",
    # Azure IoT
    "azure-iot-device>=2.13.0",
    # Azure Functions
    "azure-functions>=1.20.0",
    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    # Data
    "pydantic>=2.7.0",
    "numpy>=1.26.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.4",
    "pyyaml>=6.0.1",
    "reportlab>=4.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.7",
    "pyright>=1.1.369",
    "ipykernel>=6.29.0",
    "notebook>=7.2.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "B"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
ignore = ["data/synthetic/**"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "api*", "ingestion*", "knowledge*", "monitoring*"]
